move custom classloader specific data to avro.
