+++ comm/src/main/scala/netty/AvroSpecificDecoder.scala
@@ -1,4 +1,6 @@
-package edu.berkeley.cs.scads.comm
+package edu.berkeley.cs
+package scads
+package comm
 package netty
 
 import org.jboss.netty._
@@ -15,42 +17,6 @@ import specific._
 
 import net.lag.logging.Logger
 
-/**
- * Special vesion of SpecificData allows the user to specify the classloader
- * instead of relying on the classloader used to load SpecificData
- */
-class CustomLoaderSpecificData(loader: ClassLoader) extends SpecificData {
-  import org.apache.avro.Schema.Type._
-
-  protected val logger = Logger()
-  
-  private val classCache = new java.util.concurrent.ConcurrentHashMap[String, Class[_]]
-  
-  val NO_CLASS = new Object(){}.getClass();
-
-  override def getClass(schema: Schema): Class[_] = {
-    logger.info("Locating class %s", schema)
-    schema.getType match {
-      case FIXED | RECORD | ENUM =>
-	val name = schema.getFullName();
-        if (name == null) return null
-        var c = classCache.get(name)
-        if (c == null) {
-          try {
-            c = loader.loadClass(getClassName(schema))
-          } catch {
-            case e: ClassNotFoundException =>
-	      logger.warning("Failed to locate class %s", name)
-	      c = NO_CLASS
-          }
-          classCache.put(name, c)
-        }
-        return if(c == NO_CLASS) null else c
-      case _ => super.getClass(schema)
-    }
-  }
-}
-
 class AvroSpecificDecoder[M <: SpecificRecord](implicit m: Manifest[M])
   extends OneToOneDecoder {
 
@@ -58,12 +24,7 @@ class AvroSpecificDecoder[M <: SpecificRecord](implicit m: Manifest[M])
 
   private val schema = msgClass.newInstance.getSchema
 
-  private val msgReader = {
-    val constructor = classOf[SpecificDatumReader[M]].getDeclaredConstructor(classOf[org.apache.avro.Schema], classOf[org.apache.avro.Schema], classOf[org.apache.avro.specific.SpecificData])
-    constructor.setAccessible(true)
-    constructor.newInstance(schema, schema, new CustomLoaderSpecificData(msgClass.getClassLoader))
-  }
-
+  private val msgReader = new SpecificDatumReader[M](schema, schema, new CustomLoaderSpecificData(msgClass.getClassLoader))
   private val logger = Logger()
 
   override def decode(ctx: ChannelHandlerContext, chan: Channel, msg: AnyRef) = msg match {
